refactor(blog): rename lib/{prisma,session,password}.ts to *.server.ts
These are server-only utilities (Prisma singleton, session helpers,
password hashing). Under the two-marker convention they get the
.server.ts extension (source-protected by the file router) but NOT
the 'use server' directive: their exports aren't designed to be
RPC-callable from client code, they're called from server actions
and route handlers. The 'use server' directive is removed from all
three files at the same time.

Renames:
  lib/prisma.ts   -> lib/prisma.server.ts
  lib/session.ts  -> lib/session.server.ts
  lib/password.ts -> lib/password.server.ts

Sixteen consumer files (server actions, queries, route handlers,
middleware, password test, copilot-instructions.md, CONVENTIONS.md)
updated to import from the new paths. session.server.ts internal
import of './prisma.ts' also updated to './prisma.server.ts'.

webjs check passes on the blog after the rename.
