+++ examples/blog/.github/copilot-instructions.md
@@ -8,7 +8,7 @@ the full hosted docs are at **https://docs.webjs.com**.
 ## Persistence + scaffold rules (non-negotiable)
 
 - **Use Prisma + SQLite for data, never JSON files.** It's already wired up
-  (`prisma/schema.prisma`, `lib/prisma.ts`, `npm run db:migrate`). For ANY
+  (`prisma/schema.prisma`, `lib/prisma.server.ts`, `npm run db:migrate`). For ANY
   data the app stores (todos, posts, messages, products, comments…),
   define a Prisma model. NEVER create `data/*.json`, `db.json`, or any
   JSON file as a fake database. NEVER use module-scope arrays / Maps as
