+++ examples/blog/modules/auth/actions/login.server.ts
@@ -1,8 +1,8 @@
 'use server';
 
-import { prisma } from '../../../lib/prisma.ts';
-import { verifyPassword } from '../../../lib/password.ts';
-import { createSession } from '../../../lib/session.ts';
+import { prisma } from '../../../lib/prisma.server.ts';
+import { verifyPassword } from '../../../lib/password.server.ts';
+import { createSession } from '../../../lib/session.server.ts';
 import { validateLogin } from '../utils/validate.ts';
 import type { ActionResult, PublicUser } from '../types.ts';
