+++ examples/blog/modules/auth/actions/logout.server.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { destroySession } from '../../../lib/session.ts';
+import { destroySession } from '../../../lib/session.server.ts';
 import type { ActionResult } from '../types.ts';
 
 export async function logout(token: string | null | undefined): Promise<ActionResult<null>> {
