+++ examples/blog/modules/auth/queries/current-user.server.ts
@@ -1,7 +1,7 @@
 'use server';
 
 import { cookies } from '@webjskit/server';
-import { getUserByToken, SESSION_COOKIE } from '../../../lib/session.ts';
+import { getUserByToken, SESSION_COOKIE } from '../../../lib/session.server.ts';
 import type { PublicUser } from '../types.ts';
 
 /**
