+++ examples/blog/modules/posts/actions/delete-post.server.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { prisma } from '../../../lib/prisma.ts';
+import { prisma } from '../../../lib/prisma.server.ts';
 import { currentUser } from '../../auth/queries/current-user.server.ts';
 import { listPosts } from '../queries/list-posts.server.ts';
 import type { ActionResult } from '../../auth/types.ts';
