+++ examples/blog/modules/posts/queries/get-post.server.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { prisma } from '../../../lib/prisma.ts';
+import { prisma } from '../../../lib/prisma.server.ts';
 import { formatPost } from '../utils/slugify.ts';
 import type { PostFormatted } from '../types.ts';
